Adding POST methods bring_up/down port. Refactoring file structure.
