+++ src/xdpd/management/plugins/rest/get-controllers.h
@@ -1,3 +1,6 @@
+#ifndef GET_CONTROLLERS_H
+#define GET_CONTROLLERS_H
+
 // Copyright (c) 2014  Barnstormer Softworks, Ltd.
 
 #include <boost/regex.hpp>
@@ -11,6 +14,7 @@ namespace server{
 
 namespace xdpd{
 namespace controllers{
+namespace get{
 	/**
 	* Human readable index
 	*/
@@ -37,5 +41,8 @@ namespace controllers{
 	void list_ports(const http::server::request &, http::server::reply &, boost::cmatch&);
 	void port_detail(const http::server::request &, http::server::reply &, boost::cmatch&);
 
+} //namespace get
 } //namespace controllers
 } //namespace xdpd
+
+#endif /* GET_CONTROLLERS_H_ */
